requirements.txt
# Anthropic LLM — via Google Vertex AI (AnthropicVertex)
anthropic>=0.40.0

# LangChain text splitting
langchain>=0.1.16
langchain-community>=0.0.34
langchain-text-splitters>=0.1.0

# Google Cloud — Vertex AI (embeddings + Vector Search + Claude) and Drive
google-cloud-aiplatform>=1.51.0
google-api-python-client>=2.124.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0

# Document parsers
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2

# Microsoft Teams Bot Framework
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.9.4

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# Retry logic
tenacity>=8.2.3

# Scheduler
apscheduler>=3.10.4

# Testing
pytest>=8.1.1
pytest-asyncio>=0.23.6
